refactor(test): Continued integration split
